move sync_tag to shared file for GCP sync + quay mirror
